Heavy changes to netcdf_writer_bigfile.f90 -> netcdf_output_processing/netcdf_writer_fast.f90. I figured out that I was doing things in rather memory intensive manner. This should have resulted in slower file writing. I have thus rewrote the script to only open the files necessary and then immediately close them. I also only have one netcdf file open at a time. This will hopefully speed up file writing since the T63 files take a pretty long time to write.
Other changes include making it so that the monthly outputs are only calculated and written out when the specified year
is hit. This should result in smaller amounts of outputs generated, faster run and transfer times and quicker netcdf generation.

Lastly, for the tests of the LAI seasonal cycle, I changed it from 2.5 * Vmax to only 2.0 and also now during the normal growth period
the Vmax is 1* rather than 0.5*. The idea is to also improve the NEP seasonal cycle (as compared to CarbonTracker) while retaining
the improvements in the LAI seasonal cycle.