This is the code I will be using for FireMIP (not finalized).
I removed the LAI seasonality stuff in PHTSYN3, phenology and ctem_params (lfespany and tanhq10)

I added in the functionality to read in time-evolving lightning. The FireMIP lightning is
already cloud-to-ground so I commented out our conversion. Also the units are per month so I took out the
conversion from per year.

There was also a bug in nbpveg in ctems2 and ctemg2. This was fixed in a different branch that I haven't
yet merged to this one so I just did it here to prevent issues.

This also includes fire parameter changes to extnmois_veg and extnmois_duff as well as the lightning limits.