Some small code clean ups and a fix to how half-hourly and daily outputs are done.
Previously the halfhourly could be specified jhhstd, jhhendd,jhhsty,
and jhhendy which actually then outputted continously from jhhstd jhhsty to
jhhendd jhhendy. This is counter to what would be expected, or useful. Now it goes
from jhhstd to jhhendd for each year from jhhsty to jhhendy. If you want similar behaviour
to before just set the end and start days to 365 and 1 respectively.

I also make it so the mosaic level output files are output if the model
is run in mosaic mode or not since the per PFT level info is useful.