MAKE CONSISTENT THE TREATMENT OF INPUT GRID-LEVEL VARS
-The input vars from the various input files and CTM file are sometimes
a grid-level value but the treatment in the code was often anticipating
a per tile value due to how the loops were. To make this simple I put these
all as per tile values. This should make it simpler to deal with them and have
less 'exceptions' needed.

-Land use change is still not entirely set up right for use with a tiled run
with competition (and possibly without).