FIX RESET PROBLEM IN CLASS36CTEM INTRODUCED RECENTLY, REDUCE REP AREA IN FIRE
-The resetting of accumulated vars in CLASS36CTEM after calling ctem was originally
right after the subroutine, I moved it after ctems2 recently as I needed some vars
to be still with values in their row form. This messed up tbaracc since it was then
not being reset as the row form was then gathered back to gat without reset. This
fixes that.

-Rep area for fire is now 500 km2 since that is the value it needs to be to prevent
more than one observed fire per rep area. Other vars will need to be changed to accomodate
but for now they are not.