FIX NETCDF PROBLEM FOR PER PFT WHEN PFT LEAVES GRIDCELL
-When a PFT is pushed out of a grid cell, it would get wild values
put in by the netcdf writer. This fixes that.

-I think this won't be able to now deal with PFTs that are not there
at the start. I think I will leave as-is for now as I think I might
have to make a netcdf writer that happens during the run (to try and
deal with the file size problems we are starting to get)