BRINGING IN UPDATES FROM DIANA
-From Diana's notes:
In recent months, modelling evidence came to light showing that the old texture-based soil albedo produced large biases in tundra areas compared with the new “colour”-based approach. It was therefore agreed among the land surface modelling group that the former option should be dropped completely. Other small fixes have also been installed, and the total list for this round of revisions is as follows:
-Remove IGRALB=0 option (changes to CLASSB, CLASSG, CLASSA and GRALB)
-Use soil colour index for albedo of rock surfaces as well as mineral soils (GRALB)
-Related to the latter - In CLASSB, line 151, change IF(SAND(I,M,1).GE.0.0) to IF(SAND(I,M,1).GE.-3.5)
-Further to the WLOST fixes – add ZPOND to the calculation of EVPMAX (changes to TSOLVE, CLASST)
-The latter removes the need for the rock evaporation fix in GRINFL – deleted
-Refinement of calculation of ALVSSN and ALIRSN in SNOALBA
-Changes to dimensions of WSNO, ASVD and ASID in CLASSG and CLASSS (NTLD -> NMOS)
-Bugfix in calculation of HTCS in TMELT